# BackEnd/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateSchemaRequest:
    properties:
      collection_name:
        type: string
      fields:
        items:
          $ref: '#/definitions/models.SchemaField'
        minItems: 1
        type: array
    required:
    - collection_name
    - fields
    type: object
...
  models.SchemaField:
    properties:
      default: {}
      description:
        type: string
      name:
        type: string
      required:
        type: boolean
      type:
        description: string, number, boolean, date, object, array
        type: string
      visibility:
        description: public, private
        type: string
    required:
    - name
    - type
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@schemacraft.com
    name: SchemaCraft Support
    url: http://www.schemacraft.com/support
  description: Dynamic Schema API Builder - Create APIs on the fly with custom schemas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SchemaCraft API
  version: "1.0"
paths:
...
  /schemas:
    get:
      description: Get all schemas for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get schemas
      tags:
      - schema
    post:
      consumes:
      - application/json
      description: Create a new schema/collection definition
      parameters:
      - description: Schema data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSchemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create schema
      tags:
      - schema
  /schemas/{id}:
    delete:
      description: Delete a schema (soft delete)
      parameters:
      - description: Schema ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete schema
      tags:
      - schema
    get:
      description: Get a specific schema by ID
      parameters:
      - description: Schema ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get schema by ID
      tags:
      - schema
  /user/api-usage:
    get:
      description: Get detailed API usage statistics
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get API usage stats
      tags:
      - user
  /user/dashboard:
    get:
      description: Get dashboard data for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get user dashboard data
      tags:
      - user
  /user/regenerate-api-key:
    post:
      description: Generate a new API key for the user
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Regenerate API key
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    description: API Key for dynamic API access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"